Factor PriorityPartitionedQueue out of modelTaskQueue.ts
